🧹 Chore: Rename non routing folders with underscore (_)
