+++ app/_services/app/namespaces.js
