+++ app/layout.test.jsx
@@ -28,7 +28,7 @@ vi.mock("@inrupt/solid-ui-react", () => ({
 }));
 
 // Mocking the getMetadata function
-vi.mock("./logic/metadata/service", () => ({
+vi.mock("./_logic/metadata/service", () => ({
     getMetadata: vi.fn().mockResolvedValue({ /* mock metadata object */ })
 }));
